Make profiles replyable
