+++ components/conversation-screen/input.tsx
@@ -9,7 +9,6 @@ import {
   Platform,
   StyleSheet,
   View,
-  useWindowDimensions,
 } from 'react-native';
 import Animated, {
   useSharedValue,
@@ -36,8 +35,6 @@ import {
 } from '../default-text';
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
 import { faPaperPlane } from '@fortawesome/free-solid-svg-icons/faPaperPlane';
-import { styles as defaultTextInputStyles } from '../default-text-input';
-import { DefaultLongTextInput } from '../default-long-text-input';
 import { isMobile } from '../../util/util';
 import { Audio } from 'expo-av';
 import { uriToBase64 } from '../../api/api';
@@ -52,6 +49,7 @@ import {
   setQuote,
   useQuote,
 } from './quote';
+import { AutoResizingTextInput } from '../auto-resizing-text-input';
 
 const haptics = () => {
   if (Platform.OS !== 'web') {
@@ -158,40 +156,6 @@ const useRecorder = () => {
   };
 
   return { startRecording, stopRecording, duration };
-};
-
-const AutoResizingTextInput = (props) => {
-  const { height } = useWindowDimensions();
-
-  return (
-    <View style={{ flex: 1, maxHeight: height / 4 }}>
-      <DefaultText
-        style={{
-          zIndex: -1,
-          flexWrap: 'wrap',
-          width: '100%',
-          minHeight: 30,
-          opacity: 0,
-          fontSize: defaultTextInputStyles.textInput.fontSize,
-          paddingTop: Platform.OS === 'web' ? 5 : 4,
-        }}
-      >
-        {props.value}
-      </DefaultText>
-      <DefaultLongTextInput
-        {...props}
-        style={{
-          ...props.style,
-          outline: 'none',
-          position: 'absolute',
-          top: Platform.OS === 'web' ? 5 : 4,
-          bottom: 0,
-          left: 0,
-          right: 0,
-        }}
-      />
-    </View>
-  );
 };
 
 const Quote = ({ quote }: { quote: QuoteType | null }) => {
