+++ components/prospect-profile-screen.tsx
@@ -49,6 +49,7 @@ import { faPills } from '@fortawesome/free-solid-svg-icons/faPills'
 import { faSmoking } from '@fortawesome/free-solid-svg-icons/faSmoking'
 import { faVenusMars } from '@fortawesome/free-solid-svg-icons/faVenusMars'
 import { faPaperPlane } from '@fortawesome/free-solid-svg-icons/faPaperPlane'
+import { faReply } from '@fortawesome/free-solid-svg-icons/faReply'
 import { faLocationDot } from '@fortawesome/free-solid-svg-icons/faLocationDot'
 import { RotateCcw, Flag, X } from "react-native-feather";
 import Reanimated, {
@@ -68,6 +69,8 @@ import { AudioPlayer } from './audio-player';
 import { EnlargeablePhoto } from './enlargeable-image';
 import { commonStyles } from '../styles';
 import { useSkipped, setSkipped } from '../hide-and-block/hide-and-block';
+import { useQuote, setQuote } from './conversation-screen/quote';
+import { AutoResizingTextInput } from './auto-resizing-text-input';
 
 const Stack = createNativeStackNavigator();
 
@@ -266,6 +269,8 @@ const FloatingSendIntroButton = ({
   photoUuid,
   photoBlurhash,
 }) => {
+  const quote = useQuote();
+
   const onPress = useCallback(() => {
     if (name === undefined) return;
 
@@ -282,7 +287,12 @@ const FloatingSendIntroButton = ({
     >
       {personUuid !== undefined && name !== undefined &&
         <FontAwesomeIcon
-          icon={faPaperPlane}
+          icon={
+            // TODO: Fade between icon changes
+            quote
+              ? faReply
+              : faPaperPlane
+          }
           size={24}
           style={{
             color: 'white',
@@ -667,6 +677,7 @@ const CurriedContent = ({navigationRef, navigation, route}) => {
   useEffect(() => {
     setData(undefined);
     setNotFound(false);
+
     (async () => {
       setSkipped(personUuid, { networkState: 'fetching' });
       const response = await api('get', `/prospect-profile/${personUuid}`);
@@ -681,6 +692,10 @@ const CurriedContent = ({navigationRef, navigation, route}) => {
       );
       route.params.personId = response?.json?.person_id;
     })();
+
+    // Clear the quote, in case the user selected some text on the prospect's
+    // profile.
+    return () => setQuote(null);
   }, [personUuid]);
 
   const photoUuid = data === undefined ?
@@ -1177,9 +1192,28 @@ const Body = ({
         {!!data?.name && !!data?.about && data.about.trim() &&
           <>
             <Title style={{color: data?.theme?.title_color}}>About {data.name}</Title>
-            <DefaultText style={{color: data?.theme?.body_color}} selectable={true}>
-              {data.about}
-            </DefaultText>
+            <AutoResizingTextInput
+              style={{
+                color: data?.theme?.body_color,
+                backgroundColor: 'transparent',
+                borderWidth: 0,
+                padding: 0,
+                fontSize: undefined,
+                borderRadius: 0,
+              }}
+              value={data.about}
+              editable={Platform.OS === 'android'}
+              onSelectionChange={({ nativeEvent: { selection: { start, end } } }) => {
+                if (start === end) {
+                  setQuote(null);
+                } else {
+                  setQuote({
+                    text: data.about.slice(start, end),
+                    attribution: data.name,
+                  });
+                }
+              }}
+            />
           </>
         }
